SC2206 avoid
You are expanding a variable unquoted in an array. This will invoke the shell's sloppy word splitting and glob expansion.

Instead, prefer explicitly splitting (or not splitting):

If the variable should become a single array element, quote it.
If you want to split into lines or words, use mapfile, read -ra and/or while loops as appropriate.
This prevents the shell from doing unwanted splitting and glob expansion, and therefore avoiding problems with data containing spaces or special characters.